game: Move game initialization logic to game package
