+++ cmd/main/main.go
@@ -3,37 +3,28 @@ package main
 import (
 	"log"
 
-	"github.com/hajimehoshi/ebiten/v2"
-	"github.com/hajimehoshi/ebiten/v2/ebitenutil"
-
 	"celestial-odyssey/internal/game"
-	"celestial-odyssey/world/entities"
+	"github.com/hajimehoshi/ebiten/v2"
+)
+
+const (
+	windowWidth  = 320
+	windowHeight = 200
+	scale        = 3.0
 )
 
 func main() {
-	playerImage := loadPlayerImage()
-	player := entities.NewPlayer(100, 100, 0, playerImage)
+	setWindow()
 
-	gameInstance := game.NewGame(player)
-
-	screenWidth, screenHeight := 320, 200
-	scale := 3.0
-
-	ebiten.SetWindowTitle("Celestial Odyssey")
-	ebiten.SetWindowSize(int(float64(screenWidth)*scale), int(float64(screenHeight)*scale))
-	ebiten.SetWindowResizingMode(ebiten.WindowResizingModeEnabled)
-	ebiten.SetScreenClearedEveryFrame(true)
-
+	gameInstance := game.NewGame(windowWidth, windowHeight)
 	if err := ebiten.RunGame(gameInstance); err != nil {
 		log.Fatal(err)
 	}
 }
 
-func loadPlayerImage() *ebiten.Image {
-	img, _, err := ebitenutil.NewImageFromFile("assets/images/player.png")
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	return img
+func setWindow() {
+	ebiten.SetWindowTitle("Celestial Odyssey")
+	ebiten.SetWindowSize(int(float64(windowWidth)*scale), int(float64(windowHeight)*scale))
+	ebiten.SetWindowResizingMode(ebiten.WindowResizingModeEnabled)
+	ebiten.SetScreenClearedEveryFrame(true)
 }
